Remove Block.BASE_SIZE
I don't really think I need to be multiplying by 1 all over the place.
I have dynamic scaling based on the viewport size by now, and if I ever
wanted a pixel size for tiles again, I don't think I would need to
use it in so many places. It doesn't need to be part of the logic.
Also it shouldn't have ever really been strongly tied to an entity type,
and I'm about to make a base class for rectangular entities.