Fix overlapping collectables in all levels
This allows the logic of preventing playing the "eat" sfx over top of
the "win" sfx to work correctly. Before, it would detect that you hadn't
won the level, for all but one of the collectables in the pile, and
then that you had won, only for the last one, so it would avoid playing
the "eat" sfx for the last one, but would play it anyhow for the others.

This should also fix distortion of the "eat" sfx.

I started doing this by hand and then realized there were a BUNCH of
levels with TONS of overlapping collectables, and then I quickly found
an easy way to automate deleting all the extras, by taking advantage of
the "update-level-format.js" script I developed earlier today.