WIP: try to fix navigation bugs
...I think I understand why it's going to a different level when pressing 'r' to restart.
It's returning to `editorState`, which is set when switching activity mode to "play", but not when switching between levels.
I'm not sure exactly how this should work with undoing/redoing across levels.
But I do have a fair few tests at this point, with which to semi-safely change how this works.

Probably it shouldn't use `editorState` for reloading built-in levels, only for resetting to edit mode from playtesting.
That said, there's also the case of entering edit mode from a built-in level, in which case it would need to use a/the different source of truth as well...