Fix some flaky tests
Dragging and dropping a file only works if the event listeners are
already set up. Most of the tests start with locator-based actions or
assertions that will automatically wait for something to exist, but
not so with direct actions like triggering a file drop.

I think I may have confused the problem with the level file taking time
to load and thus overwriting the changes intended to be made to the
level after it is loaded, and so I'm removing that item from the todo,
but I'm not entirely sure about this.
That todo item may or may not be related to the same flakes, but I have
enough doubt about its accuracy to remove it, I suppose.