Fix error handling not restoring state
The round trip playthrough re-saving test was flaking.

Examining the JSON diff with [this applet](https://benjamine.github.io/jsondiffpatch/index.html), I could see that the way that it was changed was that it sometimes had a base state with nothing and the first level contents were in a delta update instead of having the level properly stored in `baseState`.

I noticed that it would fail much more easily if I closed the browsers and ran the test from a cold state.

I thought it might be calling `setLevelContent` too early, while the game was still loading, and it was getting overwritten with an empty state somehow.
Then again, I noted, the `window._forTesting.loadLevelFromText` API must be already set, since it's not getting an error calling that, so the game should be at least mostly loaded.

I added logging to track it down.
There were no `clearLevel` calls, but there were a lot of `deserialize` calls coming from `renderLevelPreview` for the Level Select screen...
Of course, when writing the level select's level preview code, I had made sure the snapshotting and restoration happened synchronously, but it was still suspicious, since it's changing the global level content state at unpredictable times.

I tried reproducing the test failure with the level preview code disabled, and it seemed to pass consistently.

So I added more logging, and I noted that when it was failing, serializing the playthrough was preceded by the `deserialize` call that restores the level content after rendering the level preview like it should be, but,
whilst the replace and restore `deserialize` calls were USUALLY in pairs, I saw one where it had a "2" badge next to one of the store calls, which meant it could sometimes fail to call `deserialize` to restore the content.
Looking closely at the code, I saw that indeed, there was a code path where it returns early if it errors.
And I have one level in the level select (when test levels are included at least) that is designed to cause an error.