Animate life cells
I'm trying to make it feel organic, but right now it feels
rigid still, because birth/death isn't animated at all,
and inconsistently organic, because it shifts through a square shape
in the course of the animation.
The animation isn't really helping with the game feel, when it isn't
tied to the gameplay.