Ensure gameplay test's loop is looping over some levels
This is just to be safe.
For a false positive to occur, it would have had to not just fail to
find the level buttons, but also trigger the game win screen.
That said, it wouldn't be the weirdest thing for the game win screen
to be shown if all the level buttons were accidentally hidden.
Better safe than sorry.