Add Go conversion of game logic module
I wanted to try making an optimized version of the level generator, because 1. it was crashing my browser tab pretty often, and 2. performance is a serious bottleneck when trying to generate puzzles with brute force. The more compute we can through at it, the more we can rely on heuristics that simply check how good a puzzle is rather than heuristics interspersed with generation code that try to avoid ever generating uninteresting puzzles.

I used ChatGPT with the prompt
> Convert this Typescript module to Go.
> Don't try to implement the imported code, just translate the import syntax, assuming the imported modules will be translated.
> Preserve comments, including todos, dead code (translate this to Go), and other notes.

It tried to use math.Signbit() in place of Math.sign() even though they're not equivalent, and defined some questionable API names like "game.EntitiesUnshift" which I changed. And I had to define some helpers. And it omitted a lot of the comments, which I added back manually, viewing the TS and Go code side-by-side.

Mostly it looks plausible, but this feels like too much code to really try to convert all at once hoping that it'll work in the end.
So I think I'm going to set this code aside and try implementing a Go version from first principles.