Make simplifyPlaythrough use MoveInput instead of Move
The Snake pointers inside Move structs are not valid across different Level states.
We have to call AnalyzeMoveRelative (or manually fix up the pointers) to get valid pointers,
so the cleanest thing is to use a representation that doesn't use the pointers until we're actually dealing with the Snake objects.